backend/users/views.py
from django.contrib.auth import get_user_model
from django.db import IntegrityError
from django.shortcuts import get_object_or_404
from djoser.views import UserViewSet
from rest_framework import status
from rest_framework.decorators import action
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response

from api.paginators import PageNumberLimitPagination
from users.models import Subscription
from users.serializers import SubscriptionSerializer

User = get_user_model()


class CustomUserViewSet(UserViewSet):
    pagination_class = PageNumberLimitPagination

    @action(
        methods=['GET'],
        detail=False,
        url_path='subscriptions',
        permission_classes=(IsAuthenticated,),
        pagination_class=PageNumberLimitPagination
    )
    def subscriptions(self, request):
        subscriptions = Subscription.objects.filter(user=request.user)
        self.paginate_queryset(subscriptions)
        serializer = SubscriptionSerializer(
            subscriptions,
            many=True,
            context={'request': request}
        )
        return self.get_paginated_response(serializer.data)

    @action(
        methods=['POST', 'DELETE'],
        detail=True,
        url_path='subscribe',
        permission_classes=(IsAuthenticated,)
    )
    def subscribe(self, request, id):
        author = get_object_or_404(User, id=id)

        if request.method == 'POST':
            try:
                subscription = Subscription.objects.create(
                    user=request.user, author=author
                )
                serializer = SubscriptionSerializer(
                    subscription,
                    context={'request': request}
                )
                return Response(
                    serializer.data, status=status.HTTP_201_CREATED
                )
            except IntegrityError as e:
                return Response(
                    {'errors': str(e)},
                    status=status.HTTP_400_BAD_REQUEST
                )
        subscription = get_object_or_404(
            Subscription, user=request.user, author=author
        )
        subscription.delete()
        return Response(status=status.HTTP_204_NO_CONTENT)
